Sign the transaction in the InvokeContractVerify method
